Arrows plot and general structure working correctly

+++ layers/layer_title.py
@@ -7,6 +7,6 @@ class LayerTitle(Layer):
         self.default_params = {'family':'Open Sans','fontsize':'15'}
 
     def render(self,plt):
-        plt.title(self.title,family=self.default_params['family'], fontsize=int(self.default_params['fontsize']))
+        plt.title(self.title,family=self.default_params['family'], fontsize=float(self.default_params['fontsize']))
 
 layer_switcher['title'] = LayerTitle
